feat: Integrate sync handler and execute sync requests
Sync requests are pushed onto a bounded queue. An executor service is not used
for pushing requests as we do not want to push items onto an unbounded queue.

A thread runs to take items off the thread and execute them.

A new context object wraps handlers to avoid having to pass them into
request constructors and helps avoid circular dependencies with the
RequestMerger, SyncHandler, and ShadowManager.

If the MQTT connection is dropped, syncing stops until the
connection is resumed.