fix: Check local and cloud to determine if update is actually necessary
When a local update is made, a sync request is created to update the cloud. When
the cloud update occurs, a notification is received that there was a cloud
update. This triggers a local update for the same content.

In order to break the cycle, when executing the request, a check is made of the
local or last synced cloud to determine if the shadow already contains the
updated content.