fix: Use connection pool and explicitly open db on startup
There is an edge case that cause an existing service to be reused when loading
plugins into a running nucleus instance. The single connection was being closed
and then reused when adding the shadow manager "back" after removing the plugin.

DB is now only updated after the service moves into the "Running" state and the
connection is replaced with a ConnectionPool.

The ConnectionPool is more desirable as each IPC request is coming on a different
thread and therefore multiple threads will be using the DB concurrently.