feat: configure H2 db settings to optimize for file size (#82)
Change config of H2 to not retain data for so long, compress data, and defrag on shutdown.

H2 retains multiple values of data in order to ensure consistency and concurrency for transactions. This causes the file size to get and stay very large until the db closes. Shadow Manager doesn't close the db isn't closed until the nucleus shuts down, this can cause shadow db to use excess space.