fix: When pushing a sync request back to the queue to retry, merge it 'underneath' any existing request
When multiple sync requests for the same shadow are made, they are merged
together in the queue.

When taking sync request 'A' from the queue to execute, a new request
'B' can be added to the queue for the same shadow. If we need to retry 'A', we
need to merge 'B' on top of 'A' rather than the normal behavior of merging the
new request on top of the existing request. In this case merging 'A' on top of
'B' will overwrite 'B' with old information.

This change allows the queue to add a request back and to treat it as an 'old'
value when merging.