fix: ignore delta field after getting the cloud document. (#91)
* fix: ignore delta field after getting the cloud document.

Currently if there was a difference between the reported and desired nodes in the cloud shadow document, the shadow manager reported an error. This will fix that issue by ignoring the delta node in the cloud shadow document. Since delta is just a computed field, we do not need to serialize it.

Also, added validation to shadow document deserialization.