fix: wait for executing sync request to finish before cancelling threads
Cancelling sync threads can cause database exceptions if an I/O operation is occurring when the thread is cancelled.

A semaphore is used along with the existing  flag to force the stopper to wait until the request is finished.

SyncStrategy common logic was refactored into the BaseSyncStrategy class to avoid duplication